refactor: move file detection logic to config package
- Centralized file extension detection and language recognition into the new config/detector.go file.
- Updated the config command to utilize the new methods for detecting file extensions and languages.
- Removed redundant functions from cmd/config.go to improve clarity and maintainability.
